chore: rename dev/consts to dev/internal/consts
